Page: support for auto collapse sidebar
