+++ src/Components/Common/components/Page.tsx
@@ -1,16 +1,33 @@
-import { RefObject } from "react";
+import { RefObject, useContext, useEffect } from "react";
 import PageTitle, { PageTitleProps } from "../PageTitle";
 import { classNames } from "../../../Utils/utils";
+import { SidebarShrinkContext } from "../Sidebar/Sidebar";
 
 interface PageProps extends PageTitleProps {
-  children: any;
-  options?: any;
+  children: React.ReactNode | React.ReactNode[];
+  options?: React.ReactNode | React.ReactNode[];
   className?: string;
   noImplicitPadding?: boolean;
   ref?: RefObject<HTMLDivElement>;
+  /**
+   * If true, the sidebar will be collapsed when mounted, and restored to original state when unmounted.
+   * @default false
+   **/
+  collapseSidebar?: boolean;
 }
 
 export default function Page(props: PageProps) {
+  const sidebar = useContext(SidebarShrinkContext);
+
+  useEffect(() => {
+    if (!props.collapseSidebar) return;
+
+    sidebar.setShrinked(true);
+    return () => {
+      sidebar.setShrinked(sidebar.shrinked);
+    };
+  }, [props.collapseSidebar]);
+
   let padding = "";
   if (!props.noImplicitPadding) {
     if (!props.hideBack || props.componentRight)
