+++ src/Components/Facility/FacilityHome.tsx
@@ -62,7 +62,7 @@ export const FacilityHome = (props: any) => {
   const { t } = useTranslation();
   const { facilityId } = props;
   const dispatch: any = useDispatch();
-  const [facilityData, setFacilityData] = useState<FacilityModel | any>({});
+  const [facilityData, setFacilityData] = useState<FacilityModel>({});
   const [capacityData, setCapacityData] = useState<Array<CapacityModal>>([]);
   const [doctorData, setDoctorData] = useState<Array<DoctorModal>>([]);
   const [isLoading, setIsLoading] = useState(false);
@@ -495,7 +495,7 @@ export const FacilityHome = (props: any) => {
             </div>
             <div className="mt-10 flex items-center gap-3">
               <div>
-                {facilityData.features?.some((feature: any) =>
+                {facilityData.features?.some((feature) =>
                   FACILITY_FEATURE_TYPES.some((f) => f.id === feature)
                 ) && (
                   <h1 className="text-lg font-semibold">Available features</h1>
@@ -624,7 +624,7 @@ export const FacilityHome = (props: any) => {
                 className="mt-2 flex w-full flex-row justify-center md:w-auto"
                 onClick={() => navigate(`/facility/${facilityId}/livefeed`)}
               >
-                <CareIcon className="care-l-video text-lg" />
+                <CareIcon icon="l-video" className="text-lg" />
                 <span>Live Monitoring</span>
               </ButtonV2>
               <ButtonV2
